主页item点击时间
描述: 无
性能影响: 无
测试场景: 本地

+++ app/src/main/java/com/simpure/expires/ui/CommodityFragment.kt
@@ -28,10 +28,10 @@ class CommodityFragment : Fragment() {
     }
 
     /** Creates commodity fragment for specific commodity ID  */
-    fun forCommodity(commodityId: Long): CommodityFragment {
+    fun forCommodity(commodityId: Int): CommodityFragment {
         val fragment = CommodityFragment()
         val args = Bundle()
-        args.putLong(KEY_COMMODITY_ID, commodityId)
+        args.putInt(KEY_COMMODITY_ID, commodityId)
         fragment.arguments = args
         return fragment
     }
